+++ R/triplet.R
@@ -43,7 +43,6 @@
 #' @seealso [random_pair_distance_emd()] and
 #'   [random_pair_distance_correlation()] for another measure of global
 #'   structure preservation.
-#' @md
 #' @examples
 #' iris_pca2 <- stats::prcomp(iris[, -5], rank. = 2, scale = FALSE, retx = TRUE)$x
 #' random_triplet_accuracy(iris, iris_pca2)